test(memory): place archive restoration proof in API suite
The raw purge and archive integration test depends on API persistence
owners, so run it in the package that provides those dependencies. Keep
core source-state tests independent of API imports in isolated CI.

Expect the single revision increment used to seed a missing source
ledger during legacy graph bootstrap.
